+++ include/iit/ecat/advr/mc_centAC_esc.h
@@ -375,7 +375,7 @@ public :
     /// Motor method implementation
     ///
     ///////////////////////////////////////////////////////
-    virtual uint16_t get_ESC_type() {
+    virtual uint32_t get_ESC_type() {
         if ( product_code == CENT_AC ) return CENT_AC;
         return NO_TYPE;
     }
